requirements.txt
absl-py==2.1.0
accelerate==1.4.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alpha_vantage==3.0.0
altair==5.5.0
annotated-types==0.7.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==42.0.5
cycler==0.12.1
datasets==3.3.2
decorator==5.1.1
deepface==0.0.93
dill==0.3.8
distlib==0.3.8
Django==5.1.1
django-allauth==65.5.0
djangorestframework==3.15.2
docstring_parser==0.16
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
flatbuffers==24.3.25
fonttools==4.53.1
frozendict==2.4.6
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
geocoder==1.38.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.3
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
graphviz==0.20.3
grpc-google-iam-v1==0.14.1
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
httplib2==0.22.0
huggingface-hub==0.24.7
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mtcnn==0.1.1
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
mysql-connector-python==8.3.0
mysqlclient==2.2.4
namex==0.0.8
narwhals==1.31.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
peewee==3.17.9
pika==1.3.2
pillow==10.4.0
platformdirs==4.2.0
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.0.0
psycopg2==2.9.10
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.2
pydotplus==2.0.2
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2024.2
PyYAML==6.0.2
ratelim==0.1.6
referencing==0.36.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requirements.py==1.0.1
retina-face==0.0.17
rich==13.8.1
rpds-py==0.23.1
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
setuptools==75.0.0
shapely==2.1.0rc1
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
sqlparse==0.5.1
streamlit==1.43.2
sympy==1.13.2
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.4.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.66.5
transformers==4.44.2
twilio==9.3.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.94
ultralytics-thop==2.0.6
uritemplate==4.1.1
urllib3==2.2.3
vertexai==1.71.1
virtualenv==20.25.1
watchdog==6.0.0
websocket-client==1.8.0
Werkzeug==3.0.1
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.1.3
xxhash==3.5.0
yarl==1.11.1
yfinance==0.2.54